requirements.txt
django
djangorestframework
psycopg2
model_mommy
